refactor(HotBoards): 인기 게시글 선정에 Strategy 패턴 적용
