+++ src/main/java/page/clab/api/domain/community/board/adapter/in/web/HotBoardsRetrievalController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import page.clab.api.domain.community.board.application.dto.response.BoardListResponseDto;
 import page.clab.api.domain.community.board.application.port.in.RetrieveHotBoardsUseCase;
+import page.clab.api.domain.community.board.domain.HotBoardStrategyType;
 import page.clab.api.global.common.dto.ApiResponse;
 
 import java.util.List;
@@ -24,14 +25,16 @@ public class HotBoardsRetrievalController {
     private final RetrieveHotBoardsUseCase retrieveHotBoardsUseCase;
 
     @Operation(summary = "[G] 커뮤니티 인기 게시글 목록 조회", description = "ROLE_GUEST 이상의 권한이 필요함<br>" +
-            "반응(이모지), 댓글 수를 합친 결과가 높은 순으로 size만큼 조회 가능")
+            "반응(이모지), 댓글 수를 합친 결과가 높은 순으로 size만큼 조회 가능<br>" +
+            "인기 게시글 선정 타입 설정 가능")
     @PreAuthorize("hasRole('GUEST')")
     @GetMapping("/hot")
     public ApiResponse<List<BoardListResponseDto>> retrieveHotBoards(
             @Min(message = "min.board.size", value = 0)
-            @RequestParam(defaultValue = "5") int size
+            @RequestParam(name = "size", defaultValue = "5") int size,
+            @RequestParam(name = "type") HotBoardStrategyType type
     ) {
-        List<BoardListResponseDto> boards = retrieveHotBoardsUseCase.retrieveHotBoards(size);
+        List<BoardListResponseDto> boards = retrieveHotBoardsUseCase.retrieveHotBoards(size, type);
         return ApiResponse.success(boards);
     }
 }
