+++ src/main/java/page/clab/api/domain/community/board/application/port/in/RetrieveHotBoardsUseCase.java
@@ -1,9 +1,10 @@
 package page.clab.api.domain.community.board.application.port.in;
 
 import page.clab.api.domain.community.board.application.dto.response.BoardListResponseDto;
+import page.clab.api.domain.community.board.domain.HotBoardStrategyType;
 
 import java.util.List;
 
 public interface RetrieveHotBoardsUseCase {
-    List<BoardListResponseDto> retrieveHotBoards(int size);
+    List<BoardListResponseDto> retrieveHotBoards(int size, HotBoardStrategyType type);
 }
