+++ src/main/java/page/clab/api/domain/community/board/domain/HotBoardStrategyType.java
@@ -1,0 +1,14 @@
+package page.clab.api.domain.community.board.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum HotBoardStrategyType {
+
+    DEFAULT("default", "댓글, 반응순 정렬");
+
+    private final String key;
+    private final String description;
+}
